requirements.txt
firebase-admin==6.9.0
joblib
numpy
scikit-learn
python-dotenv